+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - added arrow up or arrow down icon to cardsummary in myems-web
-- added sensor export, import and clone functions to myems-api
+- added sensor export, import and clone functions to myems-api, myems-admin
 - added gateway export, import and clone functions to myems-api
 ### Changed
 ### Fixed