+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added store export, import and clone functions to myems-api, myems-admin
 - added tenant export, import and clone functions to myems-api, myems-admin
 - added offline meter export, import and clone functions to myems-api
+- added virtual meter export, import and clone functions to myems-api
 ### Changed
 ### Fixed
 ### Removed