+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added space export, import and clone functions to myems-api
 - added command export, import and clone functions to myems-api
 - added energy flow diagram export, import and clone functions to myems-api
+- added tariff export, import and clone functions to myems-api
 ### Changed
 ### Fixed
 - fixed unkonwn meter uuid of energy flow diagram links in demo database